# API_HOTEL_LARAVEL/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Hotel",
        "description": "API para gerenciar as atividades de hoteis",
        "contact": {
            "email": "italojefer55@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/"
        }
    ],
    "paths": {
...
        "/api/hotels": {
            "get": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Lista todos os hotéis",
                "description": "Acesso permitido para administradores",
                "operationId": "651e79aa92da9a32b76621ef8e320dee",
                "responses": {
                    "200": {
                        "description": "Lista de hotéis"
                    }
                }
            },
            "post": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Cadastra um novo hotel",
                "description": "Acesso permitido para administradores",
                "operationId": "4a362f379f7f5bdeaa5a4a0377792a48",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Hotel Example"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Hotel cadastrado com sucesso"
                    },
                    "500": {
                        "description": "Falha ao cadastrar hotel"
                    }
                }
            }
        },
        "/api/hotels/{id}": {
            "get": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Exibe os detalhes de um hotel específico",
                "description": "Acesso permitido para administradores",
                "operationId": "8d5a1ce07c6fac0e3190d609919e6212",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do hotel"
                    },
                    "404": {
                        "description": "Hotel não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Atualiza um hotel",
                "description": "Acesso permitido para administradores",
                "operationId": "5acbf8efb42c96351d97ff7913a0c18c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Hotel Updated Example"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Hotel atualizado com sucesso"
                    },
                    "500": {
                        "description": "Falha ao atualizar hotel"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Exclui um hotel",
                "description": "Acesso permitido para administradores",
                "operationId": "bff8afd1a8fdd25a02c533ef24f7ed7d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hotel excluído com sucesso"
                    },
                    "500": {
                        "description": "Falha ao apagar hotel"
                    }
                }
            }
        },
...
    },
...
}